% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oVisAttack.r
\name{oVisAttack}
\alias{oVisAttack}
\title{Function to visualise effect-by-removal using a upset plot}
\usage{
oVisAttack(data, color = "cyan4", shape = 18, size = 2,
label.height.unit = 7)
}
\arguments{
\item{data}{a data frame. It contains all these columns including
'frac.disconnected', 'nodes.removed' and 'i' (number of nodes removed)}

\item{color}{the point color}

\item{shape}{the point shape}

\item{size}{the point size}

\item{label.height.unit}{the unit specifying the per-row height of the
combination matrix. By default, it is NULL. If specified (such as 8),
it will be used to decide 'combmatrix.label.height' and
'combmatrix.label.text'}
}
\value{
an object of class "ggplot"
}
\description{
\code{oVisAttack} is supposed to visualise effect-by-removal using a
upset plot. The top is the kite plot, and visualised below is the
combination matrix for nodes removed. It returns an object of class
"ggplot".
}
\note{
none
}
\examples{
\dontrun{
obj \%>\% transmute(value=nO, member=overlap) -> data
#data \%>\% arrange(-value) \%>\% mutate(member=fct_inorder(member)) -> data
data \%>\% oVisAttack() -> gp
gp + geom_point(aes(fill=as.factor(i)), shape=22, size=2,
color="transparent") + geom_line(aes(group=i,color=as.factor(i)))

# customised levels for members
data \%>\% separate_rows(member,sep=',') \%>\% arrange(member) \%>\%
pull(member) \%>\% unique() -> levels
data \%>\%
oVisAttack(member.levels='customised',levels.customised=levels)
}
}
\seealso{
\code{\link{oVisAttack}}
}
